move added feature to new class
