+++ runner-starter/src/main/resources/application.yaml
@@ -22,5 +22,5 @@ cache.config:
 socket:
   file: ${APISIX_LISTEN_ADDRESS}
 
-plugin-directory-absolute-path:                    #please include a '/' at the beginning and end of path
+plugin-directory-absolute-path:     #please include a '/' at the beginning and end of path
 filter-package-name:
